pyproject.toml
[project]
name = "rag-resume-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.0",
    "google-generativeai>=0.3.0",
    "llama-index>=0.14.5",
    "llama-index-embeddings-google-genai>=0.3.1",
    "llama-index-llms-google-genai>=0.6.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]
